Update ReadMe.md
Update to go public, Bored of solo coding and ready to show the world how to do this properly. I hope it goes somewhere in many moons, I happen to write everything backwards and typically bug test after thousands of lines of new updates. I will have to backdate some code on recent updates to present a working module if anyone decides to try it, report the issues and ill fi it when possible if you do not present the fix yourself to me please be sure to commit a working correction or a better way to handle a function or value.